Move OptimizerConfig to planner package
This class is related to the overall optimizer, not
just to the cost-based optimizations
