+++ core/trino-main/src/main/java/io/trino/server/ServerMainModule.java
@@ -33,7 +33,7 @@ import io.trino.block.BlockJsonSerde;
 import io.trino.client.NodeVersion;
 import io.trino.connector.ConnectorManager;
 import io.trino.connector.system.SystemConnectorModule;
-import io.trino.cost.OptimizerConfig;
+import io.trino.sql.planner.OptimizerConfig;
 import io.trino.dispatcher.DispatchManager;
 import io.trino.event.SplitMonitor;
 import io.trino.execution.DynamicFilterConfig;
