+++ core/trino-main/src/main/java/io/trino/sql/planner/OptimizerConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.cost;
+package io.trino.sql.planner;
 
 import io.airlift.configuration.Config;
 import io.airlift.configuration.ConfigDescription;
