+++ core/trino-main/src/test/java/io/trino/sql/query/TestFilterHideInacessibleColumnsSession.java
@@ -15,7 +15,7 @@ package io.trino.sql.query;
 
 import io.trino.FeaturesConfig;
 import io.trino.SystemSessionProperties;
-import io.trino.cost.OptimizerConfig;
+import io.trino.sql.planner.OptimizerConfig;
 import io.trino.execution.DynamicFilterConfig;
 import io.trino.execution.QueryManagerConfig;
 import io.trino.execution.TaskManagerConfig;
