+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/optimizer/TestHivePlans.java
@@ -16,8 +16,8 @@ package io.trino.plugin.hive.optimizer;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.io.Files;
 import io.trino.Session;
-import io.trino.cost.OptimizerConfig.JoinDistributionType;
-import io.trino.cost.OptimizerConfig.JoinReorderingStrategy;
+import io.trino.sql.planner.OptimizerConfig.JoinDistributionType;
+import io.trino.sql.planner.OptimizerConfig.JoinReorderingStrategy;
 import io.trino.plugin.hive.HdfsConfig;
 import io.trino.plugin.hive.HdfsConfiguration;
 import io.trino.plugin.hive.HdfsConfigurationInitializer;
