+++ testing/trino-benchto-benchmarks/src/test/java/io/trino/sql/planner/TestTpcdsCostBasedPlan.java
@@ -16,8 +16,8 @@ package io.trino.sql.planner;
 
 import com.google.common.collect.ImmutableMap;
 import io.trino.Session;
-import io.trino.cost.OptimizerConfig.JoinDistributionType;
-import io.trino.cost.OptimizerConfig.JoinReorderingStrategy;
+import io.trino.sql.planner.OptimizerConfig.JoinDistributionType;
+import io.trino.sql.planner.OptimizerConfig.JoinReorderingStrategy;
 import io.trino.plugin.tpcds.TpcdsConnectorFactory;
 import io.trino.testing.LocalQueryRunner;
